Update main.go
Returned this back to the single threaded edition. The multi threaded edition that splits the password wordlist file in half seems to not function properly unless the string is matched in both halves versus just one. I haven't figured it out yet and honestly the performance difference is negligible for this proof of concept learning